Add files via upload
these files work together to make a game timer in the john lemon game. the GameEnding script has been updated to make a timer that appears in the game and functions as a regular timer. if john lemon doesn't leave the house by the time the timer runs out, a game over screen will appear.